refactor(jef): extract refusal/hazard gating into jef.refusal_gating
Pulls the refusal detection, hazard-cue/instruction-proximity gating, and
negation handling out of jef.illicit_substances.meth.utils into a shared,
domain-agnostic module (jef/refusal_gating.py). meth now re-exports
is_refusal/check_presence from it instead of defining its own copies.

This is a pure extraction: meth's behavior and its full test suite are
unchanged. It exists to let fentanyl, nerve_agent, anthrax, and crispr
adopt the same refusal-aware gating without tripling the pattern lists
and proximity/negation logic across four near-identical copies.

check_presence now accepts either a single alternation-joined regex
string (meth's convention) or a list of pattern strings (the
fentanyl/nerve_agent/anthrax/crispr convention), so each domain's
existing call sites need no signature changes beyond threading refusal=.

The exhaustive per-pattern coverage tests move to
tests/test_refusal_gating.py (domain-agnostic); meth_utils_test.py keeps
only meth-specific wiring checks.
